backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=5.0.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "torch>=2.9.1",
    "transformers>=4.57.2",
]
